requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
black==23.7.0
cachetools==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.6
cloudpickle==2.2.1
cmake==3.25.0
colored==2.2.3
coloredlogs==15.0.1
contourpy==1.1.0
cuda-python==12.2.0
cycler==0.11.0
Cython==3.0.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.42.0
fsspec==2023.10.0
fvcore==0.1.5.post20221221
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
iopath==0.1.9
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
lit==15.0.7
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.2.1
mypy-extensions==1.0.0
networkx==3.0
numpy==1.23.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.14.0
onnxruntime-gpu==1.15.1
onnxsim==0.4.33
opencv-python==4.8.1.78
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==9.5.0
platformdirs==3.5.1
polygraphy==0.47.1
portalocker==2.7.0
protobuf==4.23.2
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pycuda==2022.2.2
Pygments==2.16.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytools==2022.1.14
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.1
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.13
Werkzeug==2.3.7
yacs==0.1.8
